remove test file
